pyproject.toml
[project]
name = "buckeyebot"
version = "0.1.0"
description = "Unified SMS-based AI agent for Ohio State University students"
requires-python = ">=3.11"
dependencies = [
    "beeai-framework[watsonx]",
    "flask",
    "twilio",
    "httpx",
    "python-dotenv",
    "canvasapi",
]

[project.optional-dependencies]
grubhub = [
    "Appium-Python-Client",
]
buckeyelink = [
    "playwright",
]
dev = [
    "pytest",
    "pytest-asyncio",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
